Import ``GenerateJsonSchema`` from ``fastapi._compat.v2``
where it is modified since
https://github.com/fastapi/fastapi/commit/e8b98d21871f64234520c1e770f39e6bfea3b0d5#diff-1086603fdd56511aafd1d279396b142b803e48164327148e6de26cef4cdaed81
