+++ lib/galaxy/schema/generics.py
@@ -6,7 +6,6 @@ from typing import (
 )
 
 from pydantic import BaseModel
-from pydantic.json_schema import GenerateJsonSchema
 from typing_extensions import override
 
 from galaxy.schema.fields import (
@@ -40,17 +39,6 @@ class GenericModel(BaseModel):
         elif params[0] is DecodedDatabaseIdField:
             suffix = "Request"
         return suffix
-
-
-class CustomJsonSchema(GenerateJsonSchema):
-    def get_defs_ref(self, core_mode_ref):
-        full_def = super().get_defs_ref(core_mode_ref)
-        choices = self._prioritized_defsref_choices[full_def]
-        ref, mode = core_mode_ref
-        if ref in ref_to_name:
-            for i, choice in enumerate(choices):
-                choices[i] = choice.replace(choices[0], ref_to_name[ref])  # type: ignore[call-overload]
-        return full_def
 
 
 class PatchGenericPickle:
